+++ BNA_KC_Weapons/Core/CfgPatches.hpp
@@ -1,0 +1,25 @@
+class CfgPatches
+{
+    class BNA_KC_Weapons_Core
+    {
+        author = "SweMonkey and DartRuffian";
+        addonRootClass = "BNA_KC_Weapons"
+        requiredVersion = 0.1;
+        requiredAddons[] =
+        {
+            "A3_Weapons_F"
+                // Base UGL flare
+        };
+        units[] = {};
+        weapons[] = {};
+        magazines[] =
+        {
+            "BNA_KC_UGL_FlareBlue",
+            "BNA_KC_3Rnd_UGL_FlareBlue"
+        };
+        ammo[] =
+        {
+            "BNA_KC_Flare_Blue"
+        };
+    };
+};
