+++ BNA_KC_Weapons/Core/Data/Textures/UI/BNA_KC_Mag_FlareBlue_UI.paa
